Make sure add_word_spans actually works
There seems to be a bug in xsl:number in xsltproc. When using saxon it
works, so switch to saxon at least for add_word_spans.

So you no longer need exsl for tokenizing.

Also add_word_spans is split into two xsls now to separate the
functionality of renaming the word_id attribute